Fixing some bugs related to accessor cleanup and tangent generation. Changing nomenclature for the generated binary file.

+++ glTF-Toolkit.Test/GLTFMeshUtilsTests.cpp
@@ -13,6 +13,7 @@
 
 #include <GLTFMeshUtils.h>
 #include <experimental\filesystem>
+#include <regex>
 
 #include "Helpers/TestUtils.h"
 
@@ -26,21 +27,27 @@ namespace Microsoft::glTF::Toolkit::Test
 
 	TEST_CLASS(GLTFMeshUtilsTest)
 	{
+		const char* c_ReciprocatingSawJson = "Resources\\gltf\\ReciprocatingSaw\\ReciprocatingSaw.gltf";
+		const char* c_GearboxAssyJson = "Resources\\gltf\\GearboxAssy\\GearboxAssy.gltf";
+		const char* c_2CylinderEngineJson = "Resources\\gltf\\2CylinderEngine\\2CylinderEngine.gltf";
+		const char* c_AnimatedCubeJson = "Resources\\gltf\\AnimatedCube\\AnimatedCube.gltf";
 		const char* c_WaterBottleJson = "Resources\\gltf\\WaterBottle\\WaterBottle.gltf";
 		const char* c_PrimitivesJson = "Resources\\gltf\\Primitives\\Primitives.gltf";
 
-		const char* c_OutputDirectory = "C:\\Users\\mahurlim\\Desktop\\GLTFMeshUtils\\";
-		const char*& c_TestFile = c_WaterBottleJson;
+		const char* c_OutputDirectory = "C:\\Users\\Matt\\Desktop\\GLTFMeshUtils\\";
+		const char*& c_TestFile = c_PrimitivesJson;
 
 		void ExecuteTest(const char* GLTFRelPath, const char* OutputDir, const MeshOptions& Options)
 		{
 			TestUtils::LoadAndExecuteGLTFTest(GLTFRelPath, [&](const GLTFDocument& Doc, const std::string& Path)
 			{
+				static const std::regex s_DataUriRegex = std::regex(GLTFMeshUtils::s_DataUriRegex);
+
 				std::string OutputName = TestUtils::GetFilenameExt(Path.c_str());
 				std::string OutputDirectory = OutputDir + TestUtils::GetFilename(OutputName) + "\\";
 				std::string BasePath = TestUtils::GetBasePath(Path.c_str());
 
-				auto OutputDoc = GLTFMeshUtils::ProcessMeshes(TestStreamReader(Path), Doc, Options, OutputDirectory);
+				auto OutputDoc = GLTFMeshUtils::ProcessMeshes(OutputName, Doc, TestStreamReader(Path), Options, OutputDirectory);
 				
 				// Create output directory and copy files referenced by the output document.
 				create_directories(OutputDirectory);
@@ -48,7 +55,7 @@ namespace Microsoft::glTF::Toolkit::Test
 				{
 					std::string FilePath = BasePath + p.uri;
 
-					if (exists(FilePath))
+					if (!std::regex_match(p.uri, s_DataUriRegex) && exists(FilePath))
 					{
 						copy_file(FilePath, OutputDirectory + p.uri, copy_options::overwrite_existing);
 					}
@@ -57,7 +64,7 @@ namespace Microsoft::glTF::Toolkit::Test
 				{
 					std::string FilePath = BasePath + p.uri;
 
-					if (exists(FilePath))
+					if (!std::regex_match(p.uri, s_DataUriRegex) && exists(FilePath))
 					{
 						copy_file(FilePath, OutputDirectory + p.uri, copy_options::overwrite_existing);
 					}
@@ -73,8 +80,8 @@ namespace Microsoft::glTF::Toolkit::Test
 			MeshOptions Options;
 			Options.Optimize = true;
 			Options.GenerateTangentSpace = true;
-			Options.AttributeFormat = AttributeFormat::Interleave;
-			Options.PrimitiveFormat = PrimitiveFormat::Preserved;
+			Options.AttributeFormat = AttributeFormat::Separate;
+			Options.PrimitiveFormat = PrimitiveFormat::Separate;
 
 			ExecuteTest(c_TestFile, c_OutputDirectory, Options);
 		}
@@ -84,8 +91,8 @@ namespace Microsoft::glTF::Toolkit::Test
 			MeshOptions Options;
 			Options.Optimize = true;
 			Options.GenerateTangentSpace = false;
-			Options.AttributeFormat = AttributeFormat::Interleave;
-			Options.PrimitiveFormat = PrimitiveFormat::Preserved;
+			Options.AttributeFormat = AttributeFormat::Separate;
+			Options.PrimitiveFormat = PrimitiveFormat::Separate;
 
 			ExecuteTest(c_TestFile, c_OutputDirectory, Options);
 		}
@@ -95,8 +102,8 @@ namespace Microsoft::glTF::Toolkit::Test
 			MeshOptions Options;
 			Options.Optimize = false;
 			Options.GenerateTangentSpace = true;
-			Options.AttributeFormat = AttributeFormat::Interleave;
-			Options.PrimitiveFormat = PrimitiveFormat::Combine;
+			Options.AttributeFormat = AttributeFormat::Separate;
+			Options.PrimitiveFormat = PrimitiveFormat::Separate;
 
 			ExecuteTest(c_TestFile, c_OutputDirectory, Options);
 		}
